requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
discord.py==1.6.0
PyNaCl==1.3.0